Made/tested LM317 circuit
The LM317 was configured for 12V input and 6V output to power the main laser driver circuit. With a high power LED as a stand in for the laser diode and current set to 125mA, the power dissipation of the LM317 was high enough for a heat sink to be necessary with a fan. This is not ideal so a 6V buck will be looked into. If that cannot be found, maybe it's worth trying other voltages and weighing the tradeoffs of heat dissipation in the control FET vs the regulator.