Constant power loop
This new circuit uses the same format of constant current loop but now there is a second control loop that can be switched on with a reed relay to incorporate power feedback of the photodiode. This is done by subtracting setpoints to acquire error and subtract the error from the current setpoint voltage.