AMAZING DEVELOPMENT!!!
All of this troubleshooting and different design things come down to a sign issue. It's the fact that the P-channel MOSFET turns off more with a higher voltage at the gate. SO the gate drive op-amp has been outputting the difference between it's inputs to drive the FET. But in order to enable this behavior, I've been using the fact that taking the feedback from the single node of the voltage drop across the shunt has a lower voltage with higher current. This works fine for constant current with this configuration but to incorporate constant power, the fact that the P-channel MOSFET needs a lower voltage to turn more on caused issues. The thing that I've thought of to solve this is simply having the gate drive op-amp output to a PNP or N-channel transistor to eliminate the sign issue. This means that a differential op-amp can be used to get the current feedback from the shunt or a TIA used to get power feedback from the photodiode. I can have a DPDT signal relay at the input of the gate drive op-amp which swaps the inverting input between TIA or diff-amp feedback and also swap the non-inverting input between power setpoint or current setpoint.